Adding whileParse and andParse
whileParse applies the same parser to every line and only succeeds if
all lines succeed. andParse takes two parsers and applies them to the
same input, only succeeding if both succeed. whileParse is used in
defining iambicPentameter to work on any number of lines, and andParse
is used to combine non-rhyme parsers with a rhyme constraing, such as
rhyming haiku aba.